Basic working version.
For now, it only accepts one single languge per search.
You can pass as many globs as you want for files.
The default behaviours is to get the best rated sub and save it with the
same name of the inputted file. It will decode the data from the OS
server using the info returned by the server.